Add image-builder config
This is based on image-configs from Collabora. It just fixes some
URLs hardcoded in the hwpack files, and adds eos-image-builder,
a fork of Collabora's /usr/sbin/image-builder.

Unfortunately we can only override image-configs locally if we
provide the whole of /usr/share/image/utils. That includes functions/*.sh
which actually comes from image-builder. Hence we duplicate that here.
It is unfortunate that image-builder doesn't have a cleaner separation
between code and data, but we can live with it for now.