image-builder: Handle master ostree URL and Echo special case
Missed in the swap from old to new image builder. The master images
should have their ostree URLs set for the production location.
Furthermore, Echo has a special case of echo as the base directory
instead of eos-$arch. Reworked the variables a bit to avoid duplicating
the server password.

[endlessm/eos-shell#1985]