image-builder: Adjust ostree branch names for eosN stable branches
Setup ostree branches to follow the new branching strategy where master
is the development branch and the stable branch names are eosN. For eos1
ARM, we need to maintain the URLs and branches that were released to
users so that they continue to get updates.

However, due to our previous usage of master/platform as the stable
branch name, we need to carefully pick the new branch names so we don't
push inintended updates to those same users. Current users are setup
to follow echo:master/odroidu2-echo and eos-i386:master/i386.

The new branch names will continue to follow the
eos-$arch:$branch/$platform-base notation. The echo repo is currently a
symlink to eos-armhf and can be split off if necessary. However, the
odroidu2-echo branch name will be avoided. It has been decided that no one
significant has the eos1 x86 product and we will push updates to the
eos-i386:master/i386 branch for continued development.

[endlessm/eos-shell#2601]