image-builder: Use new branch scheme for eos1, but keep compat
We want any new deployments of eos1 to use our current branch and URL
setup. However, in order to support existing deployments that are not
following this scheme, we want to continue to update the old ostree.
This is provided with two actions:

1. Update the master/odroidu2-echo branch by simply writing the ref of
   the just built eos1/odroidu2 branch.
2. Continue to create the echo -> eos-armhf symlink for the URL on the
   ostree server.

When eos1 is EOL, we can drop this compatibility setup.

[endlessm/eos-shell#2641]