image-builder: Use new EOSK1 gpg key for non-eos1 OSTrees
For non-eos1 builds (currently eos2 and master), sign ostree with the
new GPG key, EOSK1 (id BA02FC46). This has been shipped in the user's
keyring since approximately 2014-05-22.

For eos1, we want to continue to sign any updated ostrees with the old
key to give users the biggest time window to download an updated ostree
that contains the keyring with our new GPG key.

[endlessm/eos-shell#2854]