image-builder: Handle the lock file more cleanly
Currently the image builder locking happens by opening the lock file and
running all the commands in a subshell that inherit the open file
descriptor. After an error, a subsequent image build (e.g., for another
branch) fails immediately because the file is still locked. I suspect
this is because there's still a process running deep within the image
build that has the lock file open.

Rework the lock file handling so that only flock handles it. This uses
the -o option so that flock closes the file before running the command.
The only hard part is that it makes detecting lock failures difficult
because flock uses a return code of 1 in that case. In order to
differentiate locking failures, make the buildscript use a return code
of 2 on image build failures. This loses that return code of the
actually failing command, but we weren't using that, anyway.

[endlessm/eos-shell#2887]