image-builder: Provide split images for 2 disk setup
There's a desire to install an SD card in sqwerty to provide cheap
additional storage to supplement the main eMMC module. The additional
storage would primarily be filled with apps.

In order to support this, the single disk image file needs to be split
up into 2 image files - one for each disk. A new image file is created
for the 2nd disk while the full image file serves as a base for the 1st
disk. The partition for the 2nd disk will be mounted at
/var/endless-extra. An overlayfs mount with /var/endless is then made at
/endless rather than the current symlink from /endless to /var/endless.

Both the root and extra filesystems are mounted and then hooks are run
to migrate apps to the extra mounted directory. The mounting setup tries
to resemble a booted system by bind mounting the deployment /var into
the current deployment. The extra storage is then mounted on top of that
at /var/endless-extra. This is done to allow chroot hooks to run as if
they're in a fully deployed system.

[endlessm/eos-shell#4179]