image-builder: add images for Amlogic M201
New ARM board we're working with, based on Amlogic S805 SoC
(Meson8b family).

We can't simply append the device tree to the kernel image
like we could with Exynos4, because u-boot requires access to
the device tree.

Instead, we build both the kernel and devicetree into a FIT
image, which uboot knows how to work with.