image-builder: Add stage for checking if update needed
If nothing has changed, then we don't want to start a new ostree and
image build. This not only saves resources, but it also means that the
QA targets are unchanged as we prepare for a release.

Currently this stage does nothing, but later it will stop the build if
it does not detect that any changes to the previous build will result.

[endlessm/eos-shell#2029]