image-builder: Use cache file modification times to indicate update
Get the time before running the check_update hooks. If the modification
time for any files in the cache directory are updated by the hooks,
eob_check_update will exit unsuccessfully. If eob_check_update exits
successfully, this indicates that no update is needed and the
buildscript ends the image build.

[endlessm/eos-shell#2029]