image-builder: Cache apt package lists
Fetch the list of core packages and versions from apt and cache them.
This will be used to determine if any packages have been updated for
this image.

A helper script, core-packages, is added to build the package list. This
uses the python-apt library to simulate the package installation the
image builder does. It adds required and important packages without
dependencies like debootstrap then adds eos-core with dependencies. This
is very similar to how dependencies for bundles are determined.

The results are very close to the package list output by the image
build. The one known difference is that when a recommended package
cannot be installed due to a conflict, it's dependencies are left in the
package set. This only affects a few packages and should be good enough
to determine when the package set hasn't changed. In the worst case, it
would result in a false positive for a change and trigger an image
build. Fixing this would require a more complicated analysis with the
apt problem resolver.

[endlessm/eos-shell#2029]