image-builder: Fix error handling with split images
Discovered a peculiar bash behavior involving set -e (errexit),
conditional blocks and functions. From bash(1):

  The shell does not exit if the command that fails is part of the
  command list immediately following a while or until keyword, part of
  the test following the if or elif reserved words, part of any command
  executed in a && or || list except the command following the final &&
  or ||, any command in a pipeline but the last, or if the command's
  return value is being inverted with !.

Normally this is what we want when trying to allow certain failures
under set -e. By running the possibly failing command in an if or
following it with ||, we can capture the error and move on.

However, this works poorly when the command being executed is actually a
bash function. It seems that bash essentially turns off errexit when
executing the function in this case and the function does not exit
immediately when hitting errors. If the command was a normal executable,
then it would run in its own environment without this peculiarity.

To combat this, move the meat of the image splitting into its own helper
script. Then we can run that in its own environment that will correctly
exit immediately on errors. Back in eob_split, we can catch those errors
and continue with a warning as desired.

[endlessm/eos-shell#4364]