image-builder: Don't require /var/endless-extra mount at boot
The /endless mount unit already has the appropriate dependencies to
ensure that /var/endless-extra is mounted before it is, so we don't need
to require the /var/endless-extra mount for local filesystems. This
allows the /endless mount unit to better control potential failures in
mounting the extra storage SD card if it's not inserted or has failed.

[endlessm/eos-shell#4470]