Change personalities to use concise locales
For the image builds, we will use the posix locale strings
such as 'en' and 'es_GT' rather than 'Global' and 'Guatemala'.
For now, we still write the personality.conf file using
the old strings so as not to impact code that uses these strings.
In the future, once we switch the current uses of the personality
name to instead use the current locale everywhere in the system,
we will remove the personality name from personality.conf
(and perhaps completely remove personality.conf altogether).