image-builder: Publish image files immediately
Rather than waiting until all personalities are finished building to
start uploading to the image server, start uploading each personality's
files as soon as they're complete. This will allow people to start using
the images for testing sooner. A new publish-image customization hook
stage is added to support publishing the images prior to the full
publish stage.

Rsync with --delay-updates is used to copy the files into place so that
they stay hidden until the upload is complete. Since this will miss some
of the auxiliary files, a final rsync of the full output directory
remains in the publish stage.

[endlessm/eos-shell#4618]