image-builder: Trigger jenkins post-publishing jobs
Instead of waiting for the archiving and S3 uploading of images to be
triggered by the completed jenkins image build, trigger the jobs
directly from the build.

There are a few benefits here:

1. The S3 upload begins immediately after the image is published.
2. The archiving and cleanup on the server is spread out over time.
3. With incremental image cleanup on the server, the space requirements
   are reduced as it only balloons by one personality's image instead of
   the entire image set.

This requires that the image builders have the token for the
image-builder jenkins user in /etc/image-utils/jenkins-token. Any
relevant jenkins jobs need to give permission to the image-builder user
to trigger builds.

[endlessm/eos-shell#4618]