image-builder: Publish ostree to internal server
Instead of pushing directly to the production server, publish to the
internal server to provide a full internal ostree repository. This
allows the builder's ostree to be shallow while maintaining a full copy
of the ostree repository on our local server. The internal server will
be pushed to the production server in a separate step.

[endlessm/eos-shell#4741]