split-image: avoid growing disk1 instead of shrinking it
In recent e2fsprogs versions, the "resize2fs -P" behaviour changed.
At least in the case of base images (no apps), it is no longer telling
us that we can shrink.

https://bugs.launchpad.net/ubuntu/+source/e2fsprogs/+bug/1415077

In this case, the calculations for the new disk1 size actually result in
growing it.

resize2fs isn't happy about growing a filesystem beyond the end of the
block device, so it fails.

Avoid any filesystem resize in this situation.