Set ostree repo policy for eos and eosnonfree
The eos and eosnonfree products are handled a bit counterintuitively.
For both products, we want to continue using the eos OS name. For eos,
we will temporarily use a separate repo called eosfree-$arch. For
eosnonfree, it will continue using the eos-$arch repo so our users with
codecs continue to get updates. Later, when the OS is clean again, the
eosfree-$arch repo will go away and both will use the eos-$arch repo.

[endlessm/eos-shell#5531]