Pull latest ostree instead of creating new tree
This changes the architecture of the image builder so that it simply
pulls the latest commit from the ostree server rather than building a
new OS and using that for the deployment. This allows the ostree build
to be done independently of the image builds.

The other big advantage is that multiple image builds can be done all
using the same ostree commit. That improves the experience of on-demand
image builds rather than batching up all desired variants in one build.

[endlessm/eos-shell#4834]