Add --force mode to skip update check result
Sometimes a build might be desired even if the update check says one
isn't needed. This might happen if the previous build had failures that
weren't caught or if the update fact checking has a bug.

Add a -f|--force option to run the build regardless of the result of the
update check. The check is still performed so that the facts queried can
be used on the subsequent build, but the result of the check is ignored.

[endlessm/eos-shell#4834]