split: Remount root filesystem to fix boot checksum
Before the split disks are finalized, the filesystems are unmounted so
that they can be shrunk. This means that when the boot checksum wants to
be fixed, amlogic-fix-spl-checksum can't be accessed. Remount the root
filesystem temporarily to fix the boot checksum. This is moved into the
main split_image() function so that all mounting and loop device
handling happens in the same place.

[endlessm/eos-shell#5635]