Loop entire build over personalities not just image stage
While it seems that each build is a set of personalities, really each
personality needs to be treated as a separate build in order to be
handled intelligently. For instance, making checks about whether to
start the build needs to consider each personality, but the stage is run
on all personalities at once.

Rather than looping over personalities within the image stage only, run
the buildscript once per personality.

[endlessm/eos-shell#6073]