cache: Check per personality
Make the cache directory per personality and change the checks so that
they operate on each personality. This makes the app check actually
useful since it doesn't depend on the personalities used in the previous
build being the same. The personality list check is removed since it's
irrelevant to each personality.

[endlessm/eos-shell#6073]