ostree: Use bare local ostree repo
With the ostree builder split out, the repo here is entirely scratch for
pulling in the commit for the build. In that case, we can have it be a
bare repo matching the deployment's repo. This will allow the pull-local
to be sped up since it can hardlink the files instead of decompressing
and copying them. On the other hand, this will make the pull from the
remote server decompress objects, but since the commits are cached and
shared amongst personalities, this seems like a win.

This change won't take effect until the repo on the builder is wiped
away since the code tries to reuse an existing repo.

[endlessm/eos-shell#6076]