Convert run-build to python and handle all settings
Rewrite run-build in python and convert buildscript to be an executable
bash script rather than being sourced by run-build. All the settings
have been moved out of the eib.sh library to the ImageBuilder class in
run-build.

Right now all the settings are being output to buildscript via
environment variables and all the policy is in ImageBuilder. Mostly this
is done by converting a bunch of python variables to environment
variable equivalents with an EIB_ prefix. Some of the names are
inconsistent between python and what the rest of the builder expects,
but those will be changed.