Use ConfigParser for handling image build settings
Construt the configuration of the build by using the INI style
ConfigParser. This will later allow the policy to be moved out of the
builder and be declared through an explicit config file.

There are a couple differences from standard ConfigParser.

 * The ExtendedInterpolation is in use. This allows options to reference
   other options like variables. E.g., `foo = ${bar}` will reference an
   option in the same section, and `foo = ${othersect:bar}` will
   reference and option in another section. This should allow us to
   maintain our Endless build configs across arch/platform easier.

 * The default section is 'build' rather than 'DEFAULTS' since it looks
   a little nicer. All of the standard settings are in this section and
   can be accessed without specifying the section like `${product}`.

 * A special `items_no_default()` method is supplied to output a section
   without seeing everything from the default section merged in.

All the settings are still exported as environment variables to maintain
functionality in the bash backend.