Move all settings to configuration files
Instead of building the policy inside the builder code, move it out to
configuration files in the config directory. To provide flexibility,
configuration files can be grouped by product, arch, branch and
platform. Here's the ordering of the files:

 1. Default settings - $configdir/defaults.ini
 2. Product, branch, arch, platform and personality settings. These are
    in $configdir subdirectories. For example,
    $configdir/product/$product.ini
 3. System config settings - $sysconfdir/config.ini
 4. Local user settings - $configdir/local.ini

This should make it easier to change product settings in the future.