config: Convert apps lists to config settings
Take the values in the bundle_manifest files and convert them to INI
settings. The settings can be overlaid by allowing multiple
install_add_* and install_del_* keys in the apps section. What will
happen later is that the builder will concatenate all apps in the
install_add_* options and then remove all the apps in the install_del_*
options into a single install key. This isn't really pretty, but it
allows some flexibility.

Likewise, there are extra_add/del_* and split_add/del_* keys that
provide the same control as the manifest lists for determining if an app
should be moved to the extra partition or removed from the split image
entirely.

The actual apps hooks have not been updated in this commit.