Output merged configuration and publish it
Take the merged configuration and write it to $tmpdir/config.ini at the
beginning of the build. Later, copy it to the output directory to be
included in the publishing. Having it at $tmpdir/config.ini will also
later allow the build to access the configuration settings without
always relying on environment variables.