lib: Add method to merge add/del keys
In order to make the $prefix_add_* and $prefix_del_* options useful,
provide a method for merging them together to $prefix. If $prefix
already exists, that represents an explicit override in the
configuration and it's not changed. Otherwise, $prefix is set to the
difference of the add/del keys and the original add/del keys are deleted
to cleanup the section.