config: Move section specific options to defaults
The ostree_tmpdir and apps_contentdir options should really live in the
ostree and apps sections, respectively. This also fixes a namespace
collision as the environment variables get created as $EIB_OSTREE_TMPDIR
and $EIB_APPS_CONTENDIR. If there were options in the ostree and apps
sections by these names, they would override the ones from the build
section of the same name.