publish: Disable ssh StrictHostKeyChecking
When we add a new builder, it will not have the ssh host key for the
image or content server. It would be nice to distribute the servers'
host keys to the builders, but there doesn't appear to be a reliable way
to do that. Instead, disable StrictHostKeyChecking like jenkins does.

https://phabricator.endlessm.com/T11417