Rework hooks to be declared in configuration
Previously, disabling hooks for specific products was clumsy. In order
to have a hook not run on a specific product, you could either have it
in "all" and put something in the hook to exit early under some
condition, or copy it into each other valid product.

Instead, the configuration will be used to declare the hooks for each
group. By using merged options, it's easy to add or remove specific
hooks. Furthermore, the hooks can be controlled by all the other
configuration groups like arch/platform/local/etc. This also means
there's no accidental globbing of unwanted hooks still in the directory.

This work is ported from the ostree builder, which hasn't shown any
issues with it yet.

https://phabricator.endlessm.com/T11283