Add a way to specify dconf overrides in the image builder
This will be used to customize various aspects of the image builds
depending on personality, including background images and various other
configuration options.

The format is very simple for now; each override file must be enumerated
in the image builder configuration. Directories are not supported in
this version but could easily be added later.

Locks are implemented by enumerating the keys directly in the
configuration file.

https://phabricator.endlessm.com/T11594