ostree: Namespace non-eos branches
All of the free OS commits are going to the same repo, but the branch
name is $branch/$platform, like eos2.6/i386. That will obviously
conflict across products.

Leave the eos and eosnonfree products alone where they'll keep these
branches in their respective repos. However, for non-eos products like
eosinstaller and eosdev, namespace the branches like
os/$product/$branch/$platform. This is similar to how the flatpack
runtimes and apps are namespaced.

Later, the eos and eosnonfree products can be migrated to this scheme
with compat refs pointing to the current branches.

https://phabricator.endlessm.com/T11573