ostree: Rework server HTTP authentication
The production and demo repos are being made publicly accessible, so
there's no need to embed the HTTP authentication in the URL. On the
other hand, the dev repos are still being password protected, so master
builds should maintain the authentication. The password here has been
changed so that old installs can no longer see the dev repos.

https://phabricator.endlessm.com/T3784