content and image hooks for kalite content,
we get the content data (language packs) to the image at
/var/lib/kalite/content and we'll go in a wrapper around
kalite to import content at startup so we don't have to mess with a
(possibly changing) db layout.

Test if kalite helper is installed as a way to know if we should fetch
any content in the image building process.

The current content fetching policy is to get in order:
$lang-$country-minimal
$lang-$country
$lang-minimal
$lang
en-minimal (fallback)

we may need to special case languages that have really poor translation
levels in kalite.

https://phabricator.endlessm.com/T4489

Signed-off-by: Niv Sardi <xaiki@evilgiggle.com>