Add subscriptions apps to image builds
("Subscriptions apps" is the new term for "knowledge apps".)

In the past, we installed English knowledge apps on the
Arabic and French images for demo purposes, but at this time
let's only install subscriptions apps in the local language.

Note that we don't have encyclopedia apps built yet,
so we omit them from the list to install.

With the change from the legacy Virtual School to KA Lite,
the app itself is no longer localized, so it shows up on the
defaults.ini rather than the individual language files.

Also, take care of a few missing non-subscriptions apps.

Even though we do not currently support flatpaks with split images,
we provide the list of flatpaks that should be moved onto secondary
storage once we are building split images again.

https://phabricator.endlessm.com/T12665