ostree: Use os namespaced ostree refs
In order to have multiple OS's and runtimes in one repo, the ostree refs
are namespaced to begin with os/$product. In addition, the branch is
being moved to the end of the ref like flatpak. For example, the eos
product would use a deployed ref like os/eos/amd64/eos3. This is similar
to the flatpak runtime ref runtime/com.endlessm.Platform/x86_64/eos3.0.

To make this easier for other products to override the ref to use,
introduce the `ostree:product` config option and use it for the
eosnonfree and eosoem products.

https://phabricator.endlessm.com/T12778