Preload content from the content-storage server depending on image language
This will place the pre-loaded videos under /var/lib/kalite/PRELOADED/content
so that KA Lite will find it on first start so that it can be moved to its
right place and use it to initialize the content database accordingly.

Besides preloading the videos, the image builder will also preload the data
needed for the first run (language dependant) so that KA Lite don't take
ages (several minutes on my fast laptop) to run the first time.

https://phabricator.endlessm.com/T11598