eib_image: capture entire BIOS-boot partition
grub-install adds redundancy to core.img when writing it to disk if
there's enough space left over. This means that capturing only part of
the partition, using the un-expanded core.img as a size reference, is
not enough: it's only part of the written image, and the part that we
did copy does not have enough information to recover the image at boot
time.

Instead, just capture the whole thing.

https://phabricator.endlessm.com/T12559#250299