config: rearrange specification of extra apps
It is confusing to have the subscriptions apps in the defaults file,
since these are locale-specific.  Let's move the specification of
extra apps into the same files that call out their installation.

In the future we will want to redistribute other apps onto extra
storage to ship with the primary storage optimized for other
purposes.  (For example, some personalities only have a couple
subscriptions apps, so there is no reason to keep all the games
on the primary storage.)