Explicitly disable KALite content and app
While we are awaiting approval from Khan Academy to ship this content
(including the exercises that are included with the app itself),
let's disable at the top level so that we don't accidentally
add it with a new product configuration.

Also, since we don't know the format we will use for content
(e.g., we may transcode to WebM if we get permission),
let's remove the special cases for codec-free products for now.