split: Setup split hook environment like image
The split hook environment was a little different from the image hook
environment. The major reason was that the image hooks run on the ostree
checkout while the split hooks run on the mounted image.

Rework the environment so that it looks like the image hooks
environment. This involves redefining the OSTREE_DEPLOYMENT and
OSTREE_VAR variables, mounting the root over /sysroot, and mounting
/dev/pts. This will allow the split hooks to be created without
arbitrary changes from the image hooks.

https://phabricator.endlessm.com/T13450