Add back support for installing apps on split images
This is a hacky middle ground for where this might end up. Unlike in the
old app-manager world, there's currently no support for moving apps
between installations in flatpak. Furthermore, in the initial
implementation the apps are either on disk1 or disk2, but not both. So,
all apps are first uninstalled from disk1 then reinstalled on disk2.

In order to not rewrite the image flatpak installation hooks, they're
symlinked into the split stage to be called again. Since flatpak doesn't
yet support multiple system installation directories, /var/lib/flatpak
is temporarily changed to point to the disk2 flatpak directory.

https://phabricator.endlessm.com/T13450