lib: Allow multiple calls to make_tmp_ostree
Since the ostree chroot is created within the build's tmpdir and the
tmpdir is wiped at the beginning of the build, we can assume that if it
exists that the ostree chroot has been setup properly. This allows
adding a make_tmp_ostree call to the beginning of the image stage to
ensure that we have the temporary ostree for deploying with.

https://phabricator.endlessm.com/T5351