Move more build directories under the tmpdir
This makes cleanup easier and makes it easier to see that everything at
the top level besides tmp is something to be cached. Since the ostree
checkout directory is managed via configuration, it's still deleted at
the beginning of the build in case someone moved it to a non-tmp path.