Don't fetch from git when checking out
In the case of jenkins, we know it already has the latest code and
doesn't need to fetch again. For a user testing the builder, we can
assume that they'll pull the latest code if they want.

The positive effect here is that we don't need to try to wire up the SSH
key when pulling from git. That would be somewhat complex as the
eos-image-builder wrapper would need to make assumptions about the
sysconfdir and helpersdir locations.