Rename cache stage to check
The cache name never made sense for update checking. That implies cached
data, which is really the content hooks and directory in the image
builder. The fact that the previous build's information is stored
locally is an implementation detail. Rename the hooks and directory to
check to better indicate what the operations are for.