Run build within EOS build root
The image builder has always been reliant on tools installed on the
host. However, it's hard to control the host software, and it doesn't
offer the option to have branch specific build tools. An example of this
is the tmp_ostree hack - we want to use the ostree from the target
image's branch, but there's no way to manage that on the host.

Instead, do the entire build within an EOS build root. The
ImageBuildRoot class creates the root in /var/tmp/eos-image-builder
using debootstrap, and then the build specific directories are bind
mounted into the root. A new buildroot section is available in the
configuration for managing the repo, packages, components and other
parameters to debootstrap.