ImageBuilder: Use eib-chroot to run buildscript
Use the chroot helper script to run buiildscript so that all the nice
settings are there. This requires that the build specific environment
variables are moved to buildscript since eib-chroot would filter them
out. The baselib setting is moved into the ImageBuilder class so it
doesn't have to be set in an ad-hoc way.