Document building local images
In order to test changes to the image builder or build a locally
customized image, the image builder needs to be run locally rather than
on the Endless build hosts. Document how to do that with
eos-image-builder command line options and a customized
config/local.ini. An example local.ini.example is shipped showing the
common options that a tester may want to adjust.

https://phabricator.endlessm.com/T12943