eos-image-builder: Allow --show-config without locking
eos-image-builder always tries to take the lock before calling
run-build. However, this requires running the builder as root and means
that the very helpful --show-config option can't be used even though it
doesn't require locking.

Detect when --show-config is in effect and run the real builder prior to
taking the lock.