Check /etc/default/eos-chrome-plugin-update before trying to modify it
We only run the 50-chrome-plugins-updater-url hook for zn_CN images for
now, but we do it for amd64 and armhf, meaning that the file being updated
might not always be there, as it's only created in the OSTree builder for
ARM images. Thus, we need to check for its existence before trying to
modify it or the image builder will fail for Intel images.

https://phabricator.endlessm.com/T13597