Ensure GPG keyring available throughout build
Since the temporary GPG keyring was not guaranteed to be generated and
the API returned the path to the keyring, both python and shell
implementation were needed in case some code required it.

Instead, use fixed paths defined in the configuration and generate the
keyring before starting the real build or generating the buildroot. The
shell implementation is then dropped as ${EIB_KEYRING} can be safely
used anywhere in the real build.

https://phabricator.endlessm.com/T13638