config: Trim split images further
We apparently added about 700 MB to images from 3.0.4 to 3.0.5,
so we need to free up more space in order for the split images
to continue to fit on the 16G SD card.

Since we cannot achieve this with just a couple games, I've chosen
a couple large content apps that are not a great experience and
according to our metrics are very rarely used.

https://phabricator.endlessm.com/T13887