split: Add /sysroot bind mount last
This removes many aliased mounts from the mount table. Previously, /proc
and friends were showing up twice - once in the deployment and once in
the deployment's bind mounted /sysroot. Doing the mount last avoids
these aliases.

Note that this does not remove all the aliased mounts in the split disk
setup. The extra disk mount will always show up as a mount in both the
OS /var path and the deployment /var path.

https://phabricator.endlessm.com/T13879