eib_ostree: create repodir and its parents if missing
EIB_OSTREE_REPODIR=/var/cache/eos-image-builder/ostree/eos-amd64 (by
default). On the first run of the image builder, `ostree init`
is happy to create the repodir itself, but not the parent directory
/var/cache/eos-image-builder/ostree.

We could mess around with `dirname` here but since `ostree init` is
perfectly happy for the repodir to be an existing empty directory, let's
create it too.