Generate image manifest file
Getting information about completed image builds is not simple. The
information is scattered about, but the format is not standardized.
Begin providing a manifest file in JSON format that describes the build.
Currently this contains just the basic parameters of the build.

A new stage, eib_manifest, is added for assembling the JSON. This uses a
helper script to merge JSON fragment files output throughout the rest of
the build. A hook group, manifest, is added to plugin directly during
the stage.

https://phabricator.endlessm.com/T14223