Switch back to srcdir after chrooting
After 5f5bda3, the build was being run from the root directory of the
buildroot. That broke some previous assumptions that the build was
running from the source directory. In particular, getting the current
builder commit with git was failing.

Change back to the source directory after starting buildscript in the
buildroot to restore that assumption.

https://phabricator.endlessm.com/T14223