manifest: Add flatpak information
This adds a flatpak object with subsequent remotes, runtimes and apps
objects underneath it. To support split disk systems, another split-apps
object is added with a distinct set of apps from the single disk system.
Disabled remotes and local file:// remotes are skipped.

The information is gather using the flatpak GI bindings from python
outside of the chroot so that the tools in the OS don't need to be
relied on. That makes things tricky for the single disk querying because
the symlinks to share the flatpak and ostree repos don't resolve
correctly. A temporary directory with symlinks is used to handle this,
but it might be better to have this as a chroot hook and copy the
generated json fragment out of the chroot in a separate hook.

https://phabricator.endlessm.com/T14223