eib_image: make /sys available for grub-install
With the new grub-2.02-beta3, grub-install is failing with:

Installing for i386-pc platform.
device node not found
device node not found
device node not found
/usr/sbin/grub-install: error: cannot find a GRUB drive for /dev/loop0p3.  Check your device.map.

This is because the new grub has a better understanding of loop devices.
It tries to look them up with "udevadm info" (sysfs_partition_path())
and that fails because since it requires /sys/dev to be available.

Make /sys available in the grub installation environment to fix this.

https://phabricator.endlessm.com/T14258