Pull Endless apps from minor version ref instead of major version
When there are 2 minor versions in the same series being developed at
the same time (e.g., eos3.0 and eos3.1), pulling from the major ref
(e.g., eos3) may provide the wrong version. In order to ensure the
correct version is installed, the minor series branch is used to pull
and the major series branch is used to install.

To accomplish this, a temporary flatpak repo is created during the
content hook group. The minor version ref is pulled and then a major
version ref is created locally pointing to the intended minor version.
The actual installation then points the remotes temporarily at the
content repo for installation before cleaning things up for deployment.

A side benefit is that the apps are now cached between runs in the
content directory. This will speed up builds since everything does not
need to be downloaded, particularly when there are multiple
personalities being built that share common apps.

https://phabricator.endlessm.com/T14486