lib: Bring back eib_umount
In commit 61f50e2, the single target umount function was removed because
it wasn't being used and the implementation was buggy. Now it seems to
be needed again, so add it back, but this time without bugs. In
particular, a new mounts array is built up rather than trying to remove
an index from the middle of the current array.

https://phabricator.endlessm.com/T14486