Fix install of Endless apps using non-Endless runtimes
Some apps in the Endless repository use a runtime different than
com.endlessm.Platform. In that case, the content repo will not be able
to pull in the needed runtime since it assumes all runtimes come from
the Endless runtime remote. This breaks the coding build, which installs
gnome-builder from the Endless repos, but the apps depends on the GNOME
runtime.

Warn and skip these runtimes assuming that they'll installed via another
mechanism. This is the case for coding which runs the upstream apps hook
first to get the GNOME runtimes installed. Later this will all be
reworked so that Endless and non-Endless apps and runtimes use a unified
configuration.

https://phabricator.endlessm.com/T14582