config, hooks: Include GNOME flatpak repos and content in more images
From now on, we want to configure the flatpak remotes from GNOME in
all the new images without exception and then the Platform runtimes
in all EOS images with a personality (i.e. not in base images).

As an exception to the rule, we want to ship both the Platform and
SDK GNOME runtimes on all Endless Coding images including base ones,
for what we split the previous unique upstream hook in two, so that
we can make that exception cleanly.

https://phabricator.endlessm.com/T14666