ImageBuilder: Use /proc/self/mountinfo to find existing mounts
On some kernels, that mount target in /proc/self/mounts will show up
with "\040(deleted)" appended if the source of the mount has been
deleted. That trips up the code since it passes the path unaltered to
umount.

Use /proc/self/mountinfo instead. When the source is deleted, the source
path (field 4) rather than the target path (field 5) is suffixed with
//deleted. That allows the target to continue being used unaltered.

https://phabricator.endlessm.com/T14614