Make image and hooks chroot a real mount
Bubblewrap expects that / is a real mount and tries to make it a slave
mount to prevent further mounts from leaking out to the parent. That
wasn't happening since we were chrooting directly into the deployment
directory. Bind mount the deployment directory over itself so that it
appears as a real mount.

https://phabricator.endlessm.com/T14860