Handle pivot_root in image and split chroot hooks
Bubblewrap uses pivot_root to provide a clean enviroment for its
sandbox. Unfortunately, pivot_root requires that current root mount and
its parent mount are not shared mounts, which they are by default when
making new mounts.

To accomplish that, make the chroot root mount private and then bind
mount the chroot on top of itself. This will guarantee that both
conditions are satisfied.

See https://github.com/projectatomic/bubblewrap/issues/135 for details
and the workaround suggested in
https://gist.github.com/jlebon/fb6e7c6dcc3ce17d3e2a86f5938ec033.

https://phabricator.endlessm.com/T14860